requirements.txt
selenium
pytest
pytest-html
python-dotenv
webdriver-manager